refactor(oidc-common): extract client assertion provider interface
This is agreed pre-step for https://github.com/quarkusio/quarkus/issues/52232 to make the follow-up PR smaller.
