+++ extensions/oidc/runtime/src/main/java/io/quarkus/oidc/runtime/OidcTenantConfig.java
@@ -1378,15 +1378,15 @@ public interface OidcTenantConfig extends OidcClientCommonConfig {
 
         /**
          * Configures the `authorization_details` request parameter string fields of the authorization request.
-         * Please note that the `type` field is required. If the `type` is not configured, validation will fail.
-         * For example, if you configure the `quarkus.oidc.authentication.rar.simple.type=openid_credential`
+         * For example, if you configure the
+         * `quarkus.oidc.authentication.rar.simple.credential_configuration_id=vc-scope-mapping`
          * property, following `authorization_details` request parameter will be added to the authorization request:
          *
          * <pre>
          * {@code
          * [
          *    {
-         *       "type": "openid_credential"
+         *       "credential_configuration_id": "vc-scope-mapping"
          *    }
          * ]
          * }
